requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-rosetta==0.10.0
idna==3.6
Pillow==10.1.0
polib==1.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
openpyxl 
reportlab
qrcode
django-simple-history
hashids
djangorestframework
djangorestframework
djangorestframework-simplejwt